chore(refactor): refactor controller and middleware
- refactor code in the controllers to make it more maintainable
- review and rerun tests
- add try catch to cloudinary middleware

[Finishes 167120816]
