+++ server/controllers/usersController.js
@@ -1,11 +1,12 @@
 /* eslint-disable no-console */
 /* eslint-disable camelcase */
-import { validationResult } from 'express-validator';
 import bcrypt from 'bcrypt';
 import db from '../models/db';
 import saveNewUser from '../helpers/usersHelper';
 import payload from '../helpers/auth';
+import errorHelpers from '../helpers/errorHelper';
 
+const { authError, inputError } = errorHelpers;
 const { users } = db;
 const { payloader } = payload;
 const checkIfUserExists = req => users.find(user => user.email === req);
@@ -23,14 +24,7 @@ const userSignUp = async (req, res) => {
     email, first_name, last_name, phoneNumber, address, is_admin
   } = req.body;
   let { password } = req.body;
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).send({
-      status: 'error',
-      errorMsg: errors.array(),
-    });
-  }
-
+  inputError(req, res);
   const data = checkIfUserExists(email);
   if (data !== undefined) {
     return res.status(400).send({
@@ -56,16 +50,6 @@ const userSignUp = async (req, res) => {
   }
 };
 
-/**
- * Auth Error helper.
- * @param {object} res the response object.
- *
-*/
-
-const error = res => res.status(400).send({
-  status: 'error',
-  message: 'Invalid credentials'
-});
 
 /**
  * Login an existing User account.
@@ -78,20 +62,14 @@ const error = res => res.status(400).send({
 const userSignIn = async (req, res) => {
   const { email } = req.body;
   const { password } = req.body;
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).send({
-      status: 'error',
-      errorMsg: errors.array(),
-    });
-  }
+  inputError(req, res);
   const data = checkIfUserExists(email);
   if (data === undefined) {
-    return error(res);
+    return authError(res);
   }
   const isMatch = await bcrypt.compare(password, data.password);
   if (!isMatch) {
-    return error(res);
+    return authError(res);
   }
   try {
     return payloader(res, data, 200);
